requirements.txt
langchain==0.1.20
langchain_core==0.1.52
langchain_openai==0.1.6
langchain_pinecone==0.1.0
openai==1.28.1
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
